+++ lightwood/mixer/__init__.py
@@ -1,10 +1,10 @@
-from lightwood.model.unit import Unit
-from lightwood.model.base import BaseMixer
-from lightwood.model.neural import Neural
-from lightwood.model.lightgbm import LightGBM
-from lightwood.model.lightgbm_array import LightGBMArray
-from lightwood.model.sktime import SkTime
-from lightwood.model.regression import Regression
+from lightwood.mixer.unit import Unit
+from lightwood.mixer.base import BaseMixer
+from lightwood.mixer.neural import Neural
+from lightwood.mixer.lightgbm import LightGBM
+from lightwood.mixer.lightgbm_array import LightGBMArray
+from lightwood.mixer.sktime import SkTime
+from lightwood.mixer.regression import Regression
 
 
 __all__ = ['BaseMixer', 'Neural', 'LightGBM', 'LightGBMArray', 'Unit', 'Regression', 'SkTime']
