+++ lightwood/mixer/regression.py
@@ -7,7 +7,7 @@ from sklearn.linear_model import LinearRegression
 
 from lightwood.helpers.log import log
 from lightwood.api.dtype import dtype
-from lightwood.model import BaseMixer
+from lightwood.mixer import BaseMixer
 from lightwood.encoder.base import BaseEncoder
 from lightwood.data.encoded_ds import ConcatedEncodedDs, EncodedDs
